Add notifications plugins, rework webserver
This PR represents work trying to get Virtual DJ to pull artwork.
Originally it was a push-type architecture but that didn't really
make much sense since it wasn't very flexible. Then it changed to
a poll-type architecture over websocket. Then it changed to a
generic image request system that could be added to very easily.

In the middle of all of that, the text output and remote output
stuff was switched over to the notification system because
they literally have the same pattern.